run deepsea resnet model
