+++ DASH/src/networks/deepsea.py
@@ -2,34 +2,38 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-class DeepSEA_Original(nn.Module):
-    def __init__(self, ):
-        super(DeepSEA_Original, self).__init__()
-        self.Conv1 = nn.Conv1d(in_channels=4, out_channels=320, kernel_size=8)
-        self.Conv2 = nn.Conv1d(in_channels=320, out_channels=480, kernel_size=8)
-        self.Conv3 = nn.Conv1d(in_channels=480, out_channels=960, kernel_size=8)
-        self.Maxpool = nn.MaxPool1d(kernel_size=4, stride=4)
-        self.Drop1 = nn.Dropout(p=0.2)
-        self.Drop2 = nn.Dropout(p=0.5)
-        self.flatten = nn.Flatten()
-        self.Linear1 = nn.Linear(53*960, 925) 
-        self.Linear2 = nn.Linear(925, 919)
-        
+class DeepSEA(nn.Module):
+    def __init__(self, ks=None,ds=None):
+        super(DeepSEA, self).__init__()
+        k, d = 8 if ks is None else ks[0], 1 if ds is None else ds[0]
+        self.conv1 = nn.Conv1d(in_channels=4, out_channels=320, kernel_size=k, dilation=d, padding=k//2 * d)
+        k, d = 8 if ks is None else ks[1], 1 if ds is None else ds[1]
+        self.conv2 = nn.Conv1d(in_channels=320, out_channels=480, kernel_size=k, dilation=d, padding=k//2 * d)
+        k, d = 8 if ks is None else ks[2], 1 if ds is None else ds[2]
+        self.conv3 = nn.Conv1d(in_channels=480, out_channels=960, kernel_size=k, dilation=d, padding=k//2 * d)
+        self.maxpool = nn.MaxPool1d(kernel_size=4, stride=4)
+        self.drop1 = nn.Dropout(p=0.2)
+        self.drop2 = nn.Dropout(p=0.5)
+        self.linear1 = nn.Linear(59520, 925)
+        self.linear2 = nn.Linear(925, 919)
+
     def forward(self, input):
-        x = self.Conv1(input)
+        s = input.shape[-1]
+        x = self.conv1(input)[..., :s]
         x = F.relu(x)
-        x = self.Maxpool(x)
-        x = self.Drop1(x)
-        x = self.Conv2(x)
+        x = self.maxpool(x)
+        x = self.drop1(x)
+        s = x.shape[-1]
+        x = self.conv2(x)[..., :s]
         x = F.relu(x)
-        x = self.Maxpool(x)
-        x = self.Drop1(x)
-        x = self.Conv3(x)
+        x = self.maxpool(x)
+        x = self.drop1(x)
+        s = x.shape[-1]
+        x = self.conv3(x)[..., :s]
         x = F.relu(x)
-        x = self.Drop2(x)
-        x = self.flatten(x)
-        x = self.Linear1(x)
+        x = self.drop2(x)
+        x = x.view(x.size(0), -1)
+        x = self.linear1(x)
         x = F.relu(x)
-        x = self.Linear2(x)
-  
+        x = self.linear2(x)
         return x
