+++ DASH/src/run.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-python3 -W ignore main.py --dataset deepsea --arch deepsea --experiment_id 0 --valid_split 0
+python3 -W ignore main.py --dataset deepsea --arch wrn --experiment_id 0 --valid_split 0
 
 # speed test
 # python3 speed.py --experiment_id 0
